feat: Convert CineVision to multisrc
