+++ multisrc/overrides/dooplay/cinevision/res/mipmap-hdpi/ic_launcher.png
