+++ multisrc/overrides/dooplay/cinevision/res/mipmap-mdpi/ic_launcher.png
