+++ multisrc/overrides/dooplay/cinevision/res/mipmap-xhdpi/ic_launcher.png
