+++ multisrc/overrides/dooplay/cinevision/res/mipmap-xxhdpi/ic_launcher.png
