+++ multisrc/overrides/dooplay/cinevision/res/mipmap-xxxhdpi/ic_launcher.png
