+++ multisrc/overrides/dooplay/cinevision/src/extractors/StreamlareExtractor.kt
@@ -13,7 +13,7 @@ class StreamlareExtractor(private val client: OkHttpClient) {
         val body = "{\"id\":\"$id\"}".toRequestBody("application/json".toMediaType())
         val playlist = client.newCall(
             POST("https://sltube.org/api/video/stream/get", body = body),
-        ).execute().body.string()
+        ).execute().use { it.body.string() }
         val separator = "\"label\":\""
         return playlist.substringAfter(separator).split(separator).map {
             val quality = "$name - " + it.substringAfter(separator).substringBefore("\",")
