--- src/pt/cinevision/build.gradle
@@ -1,13 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-
-ext {
-    extName = 'CineVision'
-    pkgNameSuffix = 'pt.cinevision'
-    extClass = '.CineVision'
-    extVersionCode = 3
-    libVersion = '13'
-    containsNsfw = true
-}
-
-apply from: "$rootDir/common.gradle"
