--- src/pt/cinevision/src/eu/kanade/tachiyomi/animeextension/pt/cinevision/CVUrlActivity.kt
@@ -1,42 +1,0 @@
-package eu.kanade.tachiyomi.animeextension.pt.cinevision
-
-import android.app.Activity
-import android.content.ActivityNotFoundException
-import android.content.Intent
-import android.os.Bundle
-import android.util.Log
-import kotlin.system.exitProcess
-
-/**
- * Springboard that accepts https://cinevisionv5.online.com/serie/<slug> intents
- * and redirects them to the main Aniyomi process.
- */
-class CVUrlActivity : Activity() {
-
-    private val TAG = "CVUrlActivity"
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val pathSegments = intent?.data?.pathSegments
-        if (pathSegments != null && pathSegments.size > 1) {
-            val slug = pathSegments[1]
-            val searchQuery = CineVision.PREFIX_SEARCH + slug
-            val mainIntent = Intent().apply {
-                action = "eu.kanade.tachiyomi.ANIMESEARCH"
-                putExtra("query", searchQuery)
-                putExtra("filter", packageName)
-            }
-
-            try {
-                startActivity(mainIntent)
-            } catch (e: ActivityNotFoundException) {
-                Log.e(TAG, e.toString())
-            }
-        } else {
-            Log.e(TAG, "could not parse uri from intent $intent")
-        }
-
-        finish()
-        exitProcess(0)
-    }
-}
